refactor PostController and Post model to enhance delete functionality and improve pagination

+++ protected/controllers/PostController.php
@@ -105,12 +105,17 @@ class PostController extends Controller
 	 */
 	public function actionDelete($id)
 	{
-		$this->loadModel($id)->delete();
-
-		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
-		if(!isset($_GET['ajax']))
-			$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+		if(Yii::app()->request->isPostRequest)
+		{
+			$this->loadModel($id)->delete();
+	
+			if(!isset($_GET['ajax']))
+				$this->redirect(array('index'));
+		}
+		else
+			throw new CHttpException(400, 'Invalid request. Please do not repeat this request again.');
 	}
+	
 
 	/**
 	 * Lists all models.
@@ -142,15 +147,15 @@ class PostController extends Controller
 	 */
 	public function actionAdmin()
 	{
-		$model=new Post('search');
-		$model->unsetAttributes();  // clear any default values
+		$model = new Post('search');
 		if(isset($_GET['Post']))
-			$model->attributes=$_GET['Post'];
-
-		$this->render('admin',array(
-			'model'=>$model,
+			$model->attributes = $_GET['Post'];
+	
+		$this->render('admin', array(
+			'model' => $model,
 		));
 	}
+	
 
 	/**
 	 * Returns the data model based on the primary key given in the GET variable.
